+++ rules/ecmascript-6.js
@@ -17,11 +17,11 @@ module.exports = {
   // Enforce the spacing around the * in generator functions
   'generator-star-spacing': [2, { before: false, after: true }],
 
-  // Disallow arrow functions where a condition is expected
-  'no-arrow-condition': 2,
-
   // Disallow modifying variables of class declarations
   'no-class-assign': 2,
+
+  // Disallow arrow functions where they could be confused with comparisons
+  'no-confusing-arrow': 2,
 
   // Disallow modifying variables that are declared using const
   'no-const-assign': 2,